final/requirements.txt
Flask==3.1.0
gTTS==2.5.4
opencv_python==4.11.0.86
Pillow==11.1.0
playsound==1.3.0
protobuf==3.12.4
python-dotenv==1.0.1
